refactor(shared): created new helpers lib in shared HBL-74 #time 1h
