+++ libs/shared/models/dto/src/lib/CalendarAppointment/CalendarAppointment.spec.ts
@@ -4,12 +4,13 @@ import {
   CalendarAppointment,
   CalendarDate
 } from '@hubbl/shared/models/entities';
+import * as helpers from '@hubbl/shared/models/helpers';
 
-import * as Util from '../util';
+import CalendarDateDTO from '../CalendarDate';
 import CalendarAppointmentDTO from './CalendarAppointment';
-import { CalendarDateDTO } from '..';
 
 jest.mock('@hubbl/shared/models/entities');
+jest.mock('@hubbl/shared/models/helpers');
 
 const propCompare = (
   want: CalendarAppointment | CalendarAppointmentDTO,
@@ -74,7 +75,7 @@ describe('Event', () => {
         .spyOn(ClassValidator, 'validateOrReject')
         .mockRejectedValue({});
       const vpSpy = jest
-        .spyOn(Util, 'validationParser')
+        .spyOn(helpers, 'validationParser')
         .mockReturnValue({} as any);
 
       expect.assertions(3);
@@ -95,7 +96,7 @@ describe('Event', () => {
         .mockResolvedValueOnce()
         .mockRejectedValue({});
       const vpSpy = jest
-        .spyOn(Util, 'validationParser')
+        .spyOn(helpers, 'validationParser')
         .mockReturnValue({} as any);
       jest.spyOn(CalendarDateDTO, 'fromJson').mockResolvedValue({} as any);
 
@@ -146,7 +147,7 @@ describe('Event', () => {
       const vorSpy = jest
         .spyOn(ClassValidator, 'validateOrReject')
         .mockRejectedValue({});
-      const vpSpy = jest.spyOn(Util, 'validationParser').mockReturnValue({});
+      const vpSpy = jest.spyOn(helpers, 'validationParser').mockReturnValue({});
 
       expect.assertions(3);
