+++ libs/shared/models/dto/src/lib/EventTemplate/EventTemplate.ts
@@ -6,9 +6,14 @@ import {
 } from 'class-validator';
 
 import { EventTemplate, EventType } from '@hubbl/shared/models/entities';
+import {
+  numberError,
+  stringError,
+  validationParser
+} from '@hubbl/shared/models/helpers';
 
 import DTO from '../Base';
-import { DTOGroups, numberError, stringError, validationParser } from '../util';
+import { DTOGroups } from '../util';
 
 export default class EventTemplateDTO implements DTO<EventTemplate> {
   @IsNumber({}, { message: numberError('id') })
