+++ libs/shared/models/dto/src/lib/EventType/EventType.ts
@@ -7,16 +7,16 @@ import {
 } from 'class-validator';
 
 import { EventType } from '@hubbl/shared/models/entities';
-import { AppPalette } from '@hubbl/shared/types';
-
-import DTO from '../Base';
 import {
-  DTOGroups,
   enumError,
   numberError,
   stringError,
   validationParser
-} from '../util';
+} from '@hubbl/shared/models/helpers';
+import { AppPalette } from '@hubbl/shared/types';
+
+import DTO from '../Base';
+import { DTOGroups } from '../util';
 
 export default class EventTypeDTO implements DTO<EventType> {
   @IsNumber({}, { message: numberError('id') })
