+++ libs/shared/models/dto/src/lib/GymZone/GymZone.ts
@@ -8,17 +8,17 @@ import {
 } from 'class-validator';
 
 import { Calendar, GymZone, VirtualGym } from '@hubbl/shared/models/entities';
-import { GymZoneIntervals } from '@hubbl/shared/types';
-
-import DTO from '../Base';
 import {
   booleanError,
-  DTOGroups,
   enumError,
   numberError,
   stringError,
   validationParser
-} from '../util';
+} from '@hubbl/shared/models/helpers';
+import { GymZoneIntervals } from '@hubbl/shared/types';
+
+import DTO from '../Base';
+import { DTOGroups } from '../util';
 
 export default class GymZoneDTO implements DTO<GymZone> {
   @IsNumber(
