+++ libs/shared/models/dto/src/lib/Trainer/Trainer.spec.ts
@@ -2,9 +2,12 @@ import { compare, genSalt, hash } from 'bcrypt';
 import * as ClassValidator from 'class-validator';
 
 import { Gym, Trainer } from '@hubbl/shared/models/entities';
+import * as helpers from '@hubbl/shared/models/helpers';
 
 import * as Util from '../util';
 import TrainerDTO from './Trainer';
+
+jest.mock('@hubbl/shared/models/helpers');
 
 describe('TrainerDTO', () => {
   beforeEach(() => {
@@ -46,7 +49,7 @@ describe('TrainerDTO', () => {
       const vorSpy = jest
         .spyOn(ClassValidator, 'validateOrReject')
         .mockRejectedValue({});
-      const vpSpy = jest.spyOn(Util, 'validationParser').mockReturnValue({});
+      const vpSpy = jest.spyOn(helpers, 'validationParser').mockReturnValue({});
 
       expect.assertions(3);
 
@@ -97,7 +100,7 @@ describe('TrainerDTO', () => {
       const vorSpy = jest
         .spyOn(ClassValidator, 'validateOrReject')
         .mockRejectedValue({});
-      const vpSpy = jest.spyOn(Util, 'validationParser').mockReturnValue({});
+      const vpSpy = jest.spyOn(helpers, 'validationParser').mockReturnValue({});
 
       expect.assertions(3);
