+++ libs/shared/models/dto/src/lib/VirtualGym/VirtualGym.spec.ts
@@ -1,8 +1,11 @@
-import { VirtualGym } from '@hubbl/shared/models/entities';
 import * as ClassValidator from 'class-validator';
 
-import * as Util from '../util';
+import { VirtualGym } from '@hubbl/shared/models/entities';
+import * as helpers from '@hubbl/shared/models/helpers';
+
 import VirtualGymDTO from './VirtualGym';
+
+jest.mock('@hubbl/shared/models/helpers');
 
 const propCompare = (
   want: VirtualGym | VirtualGymDTO,
@@ -56,7 +59,7 @@ describe('VirtualGym', () => {
         .spyOn(ClassValidator, 'validateOrReject')
         .mockRejectedValue({});
       const vpSpy = jest
-        .spyOn(Util, 'validationParser')
+        .spyOn(helpers, 'validationParser')
         .mockReturnValue({} as any);
 
       expect.assertions(3);
@@ -102,7 +105,7 @@ describe('VirtualGym', () => {
       const vorSpy = jest
         .spyOn(ClassValidator, 'validateOrReject')
         .mockRejectedValue({});
-      const vpSpy = jest.spyOn(Util, 'validationParser').mockReturnValue({});
+      const vpSpy = jest.spyOn(helpers, 'validationParser').mockReturnValue({});
 
       expect.assertions(3);
