+++ libs/shared/models/helpers/tsconfig.lib.json
@@ -5,6 +5,6 @@
     "declaration": true,
     "types": []
   },
-  "include": ["**/*.ts", "../helpers/src/lib/validation-parser.spec.ts"],
+  "include": ["**/*.ts"],
   "exclude": ["**/*.spec.ts"]
 }
