+++ tsconfig.base.json
@@ -19,6 +19,9 @@
       "@hubbl/shared/models/entities": [
         "libs/shared/models/entities/src/index.ts"
       ],
+      "@hubbl/shared/models/helpers": [
+        "libs/shared/models/helpers/src/index.ts"
+      ],
       "@hubbl/shared/types": ["libs/shared/types/src/index.ts"]
     }
   },
