+++ workspace.json
@@ -10,6 +10,7 @@
     "landing-e2e": "apps/landing-e2e",
     "shared-models-dto": "libs/shared/models/dto",
     "shared-models-entities": "libs/shared/models/entities",
+    "shared-models-helpers": "libs/shared/models/helpers",
     "shared-types": "libs/shared/types"
   }
 }
